requirements.txt
alembic==1.7.3
aniso8601==9.0.1
apispec==5.1.1
bcrypt==3.2.0
Bcrypt-Flask==1.0.2
black==21.9b0
cffi==1.14.6
click==8.0.1
dataclasses==0.8
flake8==3.9.2
Flask==2.0.1
flask-apispec==0.11.0
Flask-JWT-Simple==0.0.3
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
regex==2021.9.24
six==1.16.0
SQLAlchemy==1.4.25
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
webargs==8.0.1
Werkzeug==2.0.1
zipp==3.5.0
